refactored Web3J files
