+++ seatunnel-connectors-v2/connector-web3j/src/main/java/org/apache/seatunnel/connectors/seatunnel/source/Web3jSourceFactory.java
@@ -29,7 +29,7 @@ import com.google.auto.service.AutoService;
 
 import java.io.Serializable;
 
-import static org.apache.seatunnel.connectors.seatunnel.config.Web3jConfig.URL;
+import static org.apache.seatunnel.connectors.seatunnel.config.Web3jSourceOptions.URL;
 
 @AutoService(Factory.class)
 public class Web3jSourceFactory implements TableSourceFactory {
